+++ src/DotNext.Tests/Runtime/Caching/ConcurrentCacheTests.cs
@@ -11,12 +11,13 @@ namespace DotNext.Runtime.Caching
             var cache = new ConcurrentCache<int, string>(10, CacheEvictionPolicy.LRU);
             True(cache.TryAdd(0, "0"));
             True(cache.TryAdd(1, "1"));
+            True(cache.TryAdd(2, "2"));
 
             Equal("0", cache[0]);
             Equal("1", cache[1]);
 
-            False(cache.TryGetValue(2, out _));
-            Equal(2, cache.Count);
+            False(cache.TryGetValue(3, out _));
+            Equal(3, cache.Count);
 
             True(cache.TryRemove(0, out var actual));
             Equal("0", actual);
@@ -165,6 +166,69 @@ namespace DotNext.Runtime.Caching
 
             True(cache.TryUpdate(0, 42, 0));
             Equal(42, cache[0]);
+        }
+
+        [Fact]
+        public static void CheckCapacity()
+        {
+            var cache = new ConcurrentCache<int, object>(5, CacheEvictionPolicy.LFU);
+            Equal(5, cache.Capacity);
+        }
+
+        [Theory]
+        [InlineData(true)]
+        [InlineData(false)]
+        public static void InsertionOrderLRU(bool descendingOrder)
+        {
+            var cache = new ConcurrentCache<int, double>(5, CacheEvictionPolicy.LRU);
+            True(cache.TryAdd(0, 0D));
+            True(cache.TryAdd(1, 1D));
+            True(cache.TryAdd(2, 2D));
+
+            cache.TryGetValue(0, out _);
+            Span<KeyValuePair<int, double>> snapshot = stackalloc KeyValuePair<int, double>[3];
+            Equal(3, cache.TakeSnapshot(snapshot, descendingOrder));
+
+            if (descendingOrder)
+            {
+                Equal(new(0, 0D), snapshot[0]);
+                Equal(new(2, 2D), snapshot[1]);
+                Equal(new(1, 1D), snapshot[2]);
+            }
+            else
+            {
+                Equal(new(0, 0D), snapshot[2]);
+                Equal(new(2, 2D), snapshot[1]);
+                Equal(new(1, 1D), snapshot[0]);
+            }
+        }
+
+        [Theory]
+        [InlineData(true)]
+        [InlineData(false)]
+        public static void InsertionOrderLFU(bool descendingOrder)
+        {
+            var cache = new ConcurrentCache<int, double>(5, CacheEvictionPolicy.LFU);
+            True(cache.TryAdd(0, 0D));
+            True(cache.TryAdd(1, 1D));
+            True(cache.TryAdd(2, 2D));
+
+            cache.TryGetValue(0, out _);
+            Span<KeyValuePair<int, double>> snapshot = stackalloc KeyValuePair<int, double>[3];
+            Equal(3, cache.TakeSnapshot(snapshot, descendingOrder));
+
+            if (descendingOrder)
+            {
+                Equal(new(2, 2D), snapshot[0]);
+                Equal(new(0, 0D), snapshot[1]);
+                Equal(new(1, 1D), snapshot[2]);
+            }
+            else
+            {
+                Equal(new(2, 2D), snapshot[2]);
+                Equal(new(0, 0D), snapshot[1]);
+                Equal(new(1, 1D), snapshot[0]);
+            }
         }
     }
 }