+++ src/DotNext/Runtime/Caching/ConcurrentCache.Eviction.cs
@@ -61,8 +61,10 @@ public partial class ConcurrentCache<TKey, TValue>
     {
         child.Links.Previous = parent;
 
-        if ((child.Links.Next = parent.Links.Next) is not null)
-            child.Links.Next.Links.Previous = child;
+        if ((child.Links.Next = parent.Links.Next) is KeyValuePair childNext)
+            childNext.Links.Previous = child;
+
+        parent.Links.Next = child;
     }
 
     private void AddFirst(KeyValuePair pair)
@@ -73,7 +75,7 @@ public partial class ConcurrentCache<TKey, TValue>
         }
         else
         {
-            lastPair.Links.Previous = pair;
+            firstPair.Links.Previous = pair;
             pair.Links.Next = firstPair;
             firstPair = pair;
         }